src/experiments/lang-reshuffle/requirements.txt
torch
numpy
pandas
tensorflow
pyyaml
argparse
logging
matplotlib
wget